refactor!: split monaco into library and binary change!: use anyhow to report errors feat: create TemplateData to pass to the templates
This commit contains quite some changes that I have worked during the past week. Most important aspect is that the code is now split into a library and binary, so other users can extend monaco or use the code without needing to use a configuration file.

The runner now is able to read props from an external configuration file, load iterations from a JSON file and modify the props and env of a job when it's being called as a step.